# docs/api/openapi-auth.yaml
openapi: 3.0.3
info:
  title: Auth API
  version: "1.0"
servers:
  - url: https://localhost:8080
paths:
  /auth/register:
    post:
      summary: Register new user (PATIENT or DOCTOR)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password, role, contactEmail]
              properties:
                login: { type: string, pattern: '^[a-zA-Z0-9._-]{3,64}$' }
                password: { type: string, format: password }
                role: { type: string, enum: [PATIENT, DOCTOR] }
                contactEmail: { type: string, format: email }
      responses:
        "202": { description: Accepted }
        "400": { description: Bad request }
        "409": { description: Conflict (login already exists) }
        "429": { description: Rate limited }
  /auth/resend-verification:
    post:
      summary: Resend verification e-mail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [loginOrEmail]
              properties:
                loginOrEmail: { type: string }
      responses:
        "202": { description: Accepted }
        "409": { description: Already active }
        "429": { description: Rate limited }
  /auth/verify:
    get:
      summary: Verify e-mail via token
      parameters:
        - in: query
          name: token
          required: true
          schema: { type: string }
      responses:
        "204": { description: Verified (or already verified) }
        "400": { description: Invalid token }
        "410": { description: Token expired }
  /admin/invites:
    post:
      summary: Create admin invite (ADMIN only)
      security: [{ basicAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "201": { description: Created }
        "401": { description: Unauthorized }
        "403": { description: Forbidden }
  /admin/invites/accept:
    get:
      summary: Accept admin invite
      parameters:
        - in: query
          name: token
          required: true
          schema: { type: string }
      responses:
        "204": { description: Accepted (or already accepted) }
        "400": { description: Invalid token }
        "410": { description: Invite expired }
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic